Use an mpsc channel to queue tagged lines (#27)
May resolve #22, but in practice - with a slight delay between writes that we may more commonly see - what we have currently may be sufficient.

The `printenv` example writes incredibly fast. Even if we listen to reads and queue writes, without a delay it's not much better. It's at least in order and not jumbled, but its far from interleaved any better.